+++ tests/testthat/test-dependencies.R
@@ -432,3 +432,12 @@ test_that("~/.Rprofile included in dev dependencies when config$user.profile()",
   expect_equal(deps$Package, "utils")
   expect_equal(deps$Dev, TRUE)
 })
+
+test_that("captures dependencies from Jupyter notebooks", {
+
+  path <- test_path("resources/notebook.ipynb")
+  deps <- dependencies(path)
+  expect_setequal(deps$Package, c("IRKernel", "MASS", "stats"))
+  expect_equal(deps$Source, rep(renv_path_normalize(path), 3))
+
+})